requirements.txt
aiocontextvars==0.2.2
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
contextvars==2.4
dataclasses==0.8
fastapi==0.65.1
h11==0.12.0
idna==2.10
immutables==0.15
loguru==0.5.3
numpy==1.19.5
opencv-python==4.5.2.52
pandas==1.1.5
Pillow==8.2.0
pycnnum==1.0.1
pydantic==1.8.2
pymongo==3.11.4
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.13.4
